SpriteID - Truncate printed string

+++ src/sprite.rs
@@ -170,6 +170,6 @@ impl Debug for SpriteID {
 
 impl Display for SpriteID {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{:x}", self.hash)
+        write!(f, "{:x}", self.hash as u32)
     }
 }
